Refactor aws.go and add aws plugin

+++ bootstrap/pkg/kfapp/aws/aws.go
@@ -40,6 +40,7 @@ import (
 	kftypes "github.com/kubeflow/kubeflow/bootstrap/pkg/apis/apps"
 	kfapis "github.com/kubeflow/kubeflow/bootstrap/v2/pkg/apis"
 	kfdefs "github.com/kubeflow/kubeflow/bootstrap/v2/pkg/apis/apps/kfdef/v1alpha1"
+	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -55,19 +56,22 @@ const (
 	CLUSTER_FEATURE_CONFIG_FILE = "cluster_features.yaml"
 	PATH                        = "path"
 	BASIC_AUTH_SECRET           = "kubeflow-login"
-)
 
-// The namespace for Istio
-const IstioNamespace = "istio-system"
+	// The namespace for Istio
+	IstioNamespace = "istio-system"
+
+	// Plugin parameter constants
+	AwsPluginName = kftypes.AWS
+)
 
 // Aws implements KfApp Interface
 // It includes the KsApp along with additional Aws types
 type Aws struct {
-	kfdefs.KfDef
+	kfDef     *kfdefs.KfDef
 	iamClient *iam.IAM
-	// requried when choose basic-auth
-	username        string
-	encodedPassword string
+
+	region string
+	roles  []string
 }
 
 // GetKfApp returns the aws kfapp. It's called by coordinator.GetKfApp
@@ -83,17 +87,42 @@ func GetPlatform(kfdef *kfdefs.KfDef) (kftypes.Platform, error) {
 	}
 
 	_aws := &Aws{
-		KfDef:     *kfdef,
+		kfDef:     kfdef,
 		iamClient: iam.New(sess),
 	}
 
 	return _aws, nil
 }
 
+// GetPluginSpec gets the plugin spec.
+func (aws *Aws) GetPluginSpec() (*AwsPluginSpec, error) {
+	// Not passing a pointer interface is a common cause of deserialization problems
+	awsPluginSpec := &AwsPluginSpec{}
+
+	err := aws.kfDef.GetPluginSpec(AwsPluginName, awsPluginSpec)
+
+	return awsPluginSpec, err
+}
+
 // GetK8sConfig is only used with ksonnet packageManager. NotImplemented in this version, return nil to use default config for API compatibility.
 func (aws *Aws) GetK8sConfig() (*rest.Config, *clientcmdapi.Config) {
 	return nil, nil
 }
+
+//if kfdef.Spec.Platform == kftypes.AWS {
+//	if options[string(kftypes.REGION)] != nil && options[string(kftypes.REGION)].(string) != "" {
+//	kfdef.Spec.Region = options[string(kftypes.REGION)].(string)
+//	} else {
+//	log.Warnf("KfDef.Spec.Region is not set; use default region %v", kftypes.DefaultAwsRegion)
+//	kfdef.Spec.Region = kftypes.DefaultAwsRegion
+//	}
+//
+//	if options[string(kftypes.ROLES)] != nil && options[string(kftypes.ROLES)].(string) != "" {
+//	kfdef.Spec.Roles = strings.Split(options[string(kftypes.ROLES)].(string), ",")
+//	} else {
+//	log.Warnf("KfDef.Spec.Roles is not set; Create new cluster instead.")
+//	}
+//}
 
 // TODO: Do we need to add annnotations for AWS.
 func createNamespace(k8sClientset *clientset.Clientset, namespace string) error {
@@ -144,7 +173,7 @@ func (aws *Aws) applyAWSInfra() error {
 	// 1. Create EKS cluster if there's no cluster available
 	if config["managed_cluster"] == true {
 		log.Infoln("Start to create eks cluster. Please wait for 10-15 mins...")
-		clusterConfigFile := filepath.Join(aws.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR, CLUSTER_CONFIG_FILE)
+		clusterConfigFile := filepath.Join(aws.kfDef.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR, CLUSTER_CONFIG_FILE)
 		output, err := exec.Command("eksctl", "create", "cluster", "--config-file="+clusterConfigFile).Output()
 		if err != nil {
 			return &kfapis.KfError{
@@ -167,12 +196,13 @@ func (aws *Aws) applyAWSInfra() error {
 
 		var nodeGroupIamRoles []string
 		for _, output := range listRolesOutput.Roles {
-			if strings.HasPrefix(*output.RoleName, "eksctl-"+aws.KfDef.Name+"-") && strings.Contains(*output.RoleName, "NodeInstanceRole") {
+			if strings.HasPrefix(*output.RoleName, "eksctl-"+aws.kfDef.Name+"-") && strings.Contains(*output.RoleName, "NodeInstanceRole") {
 				nodeGroupIamRoles = append(nodeGroupIamRoles, *output.RoleName)
 			}
 		}
 
-		aws.Spec.Roles = nodeGroupIamRoles
+		// TODO: any context sharing?
+		//		aws.kfDef.Spec.Roles = nodeGroupIamRoles
 		aws.writeConfigFile()
 	} else {
 		log.Infof("You already have cluster setup. Skip creating new eks cluster. ")
@@ -180,15 +210,15 @@ func (aws *Aws) applyAWSInfra() error {
 
 	// 2. Attach IAM Policies
 	// TODO: Once pod level IAM complete, we don't need worker group roles. Authorize cloud services using service account.
-	for _, iamRole := range aws.Spec.Roles {
+	for _, iamRole := range aws.roles {
 		aws.attachIamInlinePolicy(iamRole, "iam_alb_ingress_policy",
-			filepath.Join(aws.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR, "iam_alb_ingress_policy.json"))
+			filepath.Join(aws.kfDef.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR, "iam_alb_ingress_policy.json"))
 		aws.attachIamInlinePolicy(iamRole, "iam_csi_fsx_policy",
-			filepath.Join(aws.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR, "iam_csi_fsx_policy.json"))
+			filepath.Join(aws.kfDef.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR, "iam_csi_fsx_policy.json"))
 
 		if config["worker_node_group_logging"] == "true" {
 			aws.attachIamInlinePolicy(iamRole, "iam_cloudwatch_policy",
-				filepath.Join(aws.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR, "iam_cloudwatch_policy.json"))
+				filepath.Join(aws.kfDef.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR, "iam_cloudwatch_policy.json"))
 		}
 	}
 
@@ -272,8 +302,8 @@ func (aws *Aws) updateClusterConfig(clusterConfigFile string) error {
 
 	// Replace placeholder with clusterName and Region
 	metadata := res.(map[string]interface{})
-	metadata["name"] = aws.KfDef.Name
-	metadata["region"] = aws.KfDef.Spec.Region
+	metadata["name"] = aws.kfDef.Name
+	metadata["region"] = aws.region
 	data["metadata"] = metadata
 
 	if buf, err = yaml.Marshal(data); err != nil {
@@ -295,9 +325,9 @@ func (aws *Aws) updateClusterConfig(clusterConfigFile string) error {
 // ${KUBEFLOW_SRC}/${KFAPP}/aws_config -> destDir (dest)
 func (aws *Aws) generateInfraConfigs() error {
 	// 1. Copy and Paste all files from `sourceDir` to `destDir`
-	parentDir := path.Dir(aws.Spec.Repo)
+	parentDir := path.Dir(aws.kfDef.Spec.Repo)
 	sourceDir := path.Join(parentDir, "deployment/aws/infra_configs")
-	destDir := path.Join(aws.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR)
+	destDir := path.Join(aws.kfDef.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR)
 
 	if _, err := os.Stat(destDir); os.IsNotExist(err) {
 		log.Infof("Creating AWS infrastructure configs in directory %v", destDir)
@@ -328,7 +358,7 @@ func (aws *Aws) generateInfraConfigs() error {
 		}
 	}
 
-	// 2. Reading from cluster_config.yaml and replace placeholders with values in aws.spec.
+	// 2. Reading from cluster_config.yaml and replace placeholders with values in aws.kfDef.Spec.
 	clusterConfigFile := filepath.Join(destDir, CLUSTER_CONFIG_FILE)
 	if err := aws.updateClusterConfig(clusterConfigFile); err != nil {
 		return err
@@ -344,7 +374,7 @@ func (aws *Aws) generateInfraConfigs() error {
 		}
 	}
 
-	if aws.KfDef.Spec.Roles != nil && len(aws.KfDef.Spec.Roles) != 0 {
+	if aws.roles != nil && len(aws.roles) != 0 {
 		featureCfg["managed_cluster"] = false
 	} else {
 		featureCfg["managed_cluster"] = true
@@ -382,25 +412,60 @@ func insertSecret(client *clientset.Clientset, secretName string, namespace stri
 
 // Use username and password provided by user and create secret for basic auth.
 func (aws *Aws) createBasicAuthSecret(client *clientset.Clientset) error {
+	awsPluginSpec, err := aws.GetPluginSpec()
+	if err != nil {
+		return err
+	}
+
+	if awsPluginSpec.Auth == nil || awsPluginSpec.Auth.BasicAuth == nil || awsPluginSpec.Auth.BasicAuth.Password.Name == "" {
+		err := errors.WithStack(fmt.Errorf("BasicAuth.Password.Name must be set"))
+		return err
+	}
+
+	password, err := aws.kfDef.GetSecret(awsPluginSpec.Auth.BasicAuth.Password.Name)
+	if err != nil {
+		log.Errorf("There was a problem getting the password for basic auth; error %v", err)
+		return err
+	}
+
+	encodedPassword, err := base64EncryptPassword(password)
+	if err != nil {
+		log.Errorf("There was a problem encrypting the password; %v", err)
+		return err
+	}
+
 	secret := &v1.Secret{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      BASIC_AUTH_SECRET,
-			Namespace: aws.Namespace,
+			Namespace: aws.kfDef.Namespace,
 		},
 		Data: map[string][]byte{
-			"username":     []byte(aws.username),
-			"passwordhash": []byte(aws.encodedPassword),
+			"username":     []byte(awsPluginSpec.Auth.BasicAuth.Username),
+			"passwordhash": []byte(encodedPassword),
 		},
 	}
-	_, err := client.CoreV1().Secrets(aws.Namespace).Update(secret)
+	_, err = client.CoreV1().Secrets(aws.kfDef.Namespace).Update(secret)
 	if err != nil {
 		log.Warnf("Updating basic auth login failed, trying to create one: %v", err)
-		return insertSecret(client, BASIC_AUTH_SECRET, aws.Namespace, map[string][]byte{
-			"username":     []byte(aws.username),
-			"passwordhash": []byte(aws.encodedPassword),
+		return insertSecret(client, BASIC_AUTH_SECRET, aws.kfDef.Namespace, map[string][]byte{
+			"username":     []byte(awsPluginSpec.Auth.BasicAuth.Username),
+			"passwordhash": []byte(encodedPassword),
 		})
 	}
 	return nil
+}
+
+func base64EncryptPassword(password string) (string, error) {
+	passwordHash, err := bcrypt.GenerateFromPassword([]byte(password), 10)
+	if err != nil {
+		return "", &kfapis.KfError{
+			Code:    int(kfapis.INVALID_ARGUMENT),
+			Message: fmt.Sprintf("Error when hashing password: %v", err),
+		}
+	}
+	encodedPassword := base64.StdEncoding.EncodeToString(passwordHash)
+
+	return encodedPassword, nil
 }
 
 // Init initializes aws kfapp - platform
@@ -428,14 +493,14 @@ func (aws *Aws) Init(resources kftypes.ResourceEnum) error {
 	}
 
 	// Finish initialization and write spec to config file
-	swaggerFile := filepath.Join(path.Dir(aws.Spec.Repo), kftypes.DefaultSwaggerFile)
-	aws.Spec.ServerVersion = "file:" + swaggerFile
+	swaggerFile := filepath.Join(path.Dir(aws.kfDef.Spec.Repo), kftypes.DefaultSwaggerFile)
+	aws.kfDef.Spec.ServerVersion = "file:" + swaggerFile
 
 	createConfigErr := aws.writeConfigFile()
 	if createConfigErr != nil {
 		return &kfapis.KfError{
 			Code:    int(kfapis.INVALID_ARGUMENT),
-			Message: fmt.Sprintf("Cannot create config file app.yaml in %v", aws.Spec.AppDir),
+			Message: fmt.Sprintf("Cannot create config file app.yaml in %v", aws.kfDef.Spec.AppDir),
 		}
 	}
 
@@ -461,24 +526,33 @@ func (aws *Aws) Generate(resources kftypes.ResourceEnum) error {
 		}
 	}
 
+	setAwsPluginDefaultsErr := aws.setAwsPluginDefaults()
+	if setAwsPluginDefaultsErr != nil {
+		return &kfapis.KfError{
+			Code: setAwsPluginDefaultsErr.(*kfapis.KfError).Code,
+			Message: fmt.Sprintf("aws set aws plugin defaults Error %v",
+				setAwsPluginDefaultsErr.(*kfapis.KfError).Message),
+		}
+	}
+
 	// Customize parameters for AWS components
-	aws.Spec.ComponentParams["aws-alb-ingress-controller"] = setNameVal(aws.Spec.ComponentParams["aws-alb-ingress-controller"], "clusterName", aws.KfDef.Name, true)
-	aws.Spec.ComponentParams["istio-ingress"] = setNameVal(aws.Spec.ComponentParams["istio-ingress"], "namespace", IstioNamespace, false)
+	aws.kfDef.Spec.ComponentParams["aws-alb-ingress-controller"] = setNameVal(aws.kfDef.Spec.ComponentParams["aws-alb-ingress-controller"], "clusterName", aws.kfDef.Name, true)
+	aws.kfDef.Spec.ComponentParams["istio-ingress"] = setNameVal(aws.kfDef.Spec.ComponentParams["istio-ingress"], "namespace", IstioNamespace, false)
 
 	// TODO: Doesn't need mysqlPd and minioPd overlay and they bind to GCE now.
 
 	// Special handling for cloud watch logs of worker node groups
 	if awsFeatureConfig["worker_node_group_logging"] == true {
-		aws.Spec.Components = append(aws.Spec.Components, "fluentd-cloud-watch")
-		aws.Spec.ComponentParams["fluentd-cloud-watch"] = setNameVal(aws.Spec.ComponentParams["fluentd-cloud-watch"], "clusterName", aws.KfDef.Name, true)
-		aws.Spec.ComponentParams["fluentd-cloud-watch"] = setNameVal(aws.Spec.ComponentParams["fluentd-cloud-watch"], "region", aws.KfDef.Spec.Region, true)
+		aws.kfDef.Spec.Components = append(aws.kfDef.Spec.Components, "fluentd-cloud-watch")
+		aws.kfDef.Spec.ComponentParams["fluentd-cloud-watch"] = setNameVal(aws.kfDef.Spec.ComponentParams["fluentd-cloud-watch"], "clusterName", aws.kfDef.Name, true)
+		aws.kfDef.Spec.ComponentParams["fluentd-cloud-watch"] = setNameVal(aws.kfDef.Spec.ComponentParams["fluentd-cloud-watch"], "region", aws.region, true)
 	}
 
 	// Special handling for sparkakus
-	for _, component := range aws.Spec.Components {
+	for _, component := range aws.kfDef.Spec.Components {
 		if component == "spartakus" {
 			rand.Seed(time.Now().UnixNano())
-			aws.Spec.ComponentParams["spartakus"] = setNameVal(aws.Spec.ComponentParams["spartakus"],
+			aws.kfDef.Spec.ComponentParams["spartakus"] = setNameVal(aws.kfDef.Spec.ComponentParams["spartakus"],
 				"usageId", strconv.Itoa(rand.Int()), true)
 		}
 	}
@@ -487,36 +561,45 @@ func (aws *Aws) Generate(resources kftypes.ResourceEnum) error {
 	if createConfigErr != nil {
 		return &kfapis.KfError{
 			Code: createConfigErr.(*kfapis.KfError).Code,
-			Message: fmt.Sprintf("Cannot create config file app.yaml in %v: %v", aws.Spec.AppDir,
+			Message: fmt.Sprintf("Cannot create config file app.yaml in %v: %v", aws.kfDef.Spec.AppDir,
 				createConfigErr.(*kfapis.KfError).Message),
 		}
 	}
 	return nil
 }
 
+func (aws *Aws) setAwsPluginDefaults() error {
+	awsPluginSpec, err := aws.GetPluginSpec()
+
+	if err != nil {
+		return err
+	}
+
+	if isValid, msg := awsPluginSpec.IsValid(); !isValid {
+		log.Errorf("AwsPluginSpec isn't valid; error %v", msg)
+		return fmt.Errorf(msg)
+	}
+
+	aws.region = awsPluginSpec.Region
+	aws.roles = awsPluginSpec.Roles
+
+	return nil
+}
+
 // Remind: Need to be thread-safe: this entry is share among kfctl and deploy app
 func (aws *Aws) Apply(resources kftypes.ResourceEnum) error {
 	// kfctl only
-	// TODO: Enable BasicAuth later
-	if aws.Spec.UseBasicAuth {
-		if os.Getenv(kftypes.KUBEFLOW_USERNAME) == "" || os.Getenv(kftypes.KUBEFLOW_PASSWORD) == "" {
-			return &kfapis.KfError{
-				Code: int(kfapis.INVALID_ARGUMENT),
-				Message: fmt.Sprintf("aws apply needs ENV %v and %v set when using basic auth",
-					kftypes.KUBEFLOW_USERNAME, kftypes.KUBEFLOW_PASSWORD),
-			}
+	setAwsPluginDefaultsErr := aws.setAwsPluginDefaults()
+	if setAwsPluginDefaultsErr != nil {
+		return &kfapis.KfError{
+			Code: setAwsPluginDefaultsErr.(*kfapis.KfError).Code,
+			Message: fmt.Sprintf("aws set aws plugin defaults Error %v",
+				setAwsPluginDefaultsErr.(*kfapis.KfError).Message),
 		}
+	}
 
-		aws.username = os.Getenv(kftypes.KUBEFLOW_USERNAME)
-		password := os.Getenv(kftypes.KUBEFLOW_PASSWORD)
-		passwordHash, err := bcrypt.GenerateFromPassword([]byte(password), 10)
-		if err != nil {
-			return &kfapis.KfError{
-				Code:    int(kfapis.INVALID_ARGUMENT),
-				Message: fmt.Sprintf("Error when hashing password: %v", err),
-			}
-		}
-		aws.encodedPassword = base64.StdEncoding.EncodeToString(passwordHash)
+	if aws.kfDef.Spec.UseBasicAuth {
+		// TODO: // TODO: Enable BasicAuth later
 	} else {
 		// We should force people either use OIDC or Coginito if user doesn't enable BASIC_AUTHN
 	}
@@ -535,6 +618,15 @@ func (aws *Aws) Apply(resources kftypes.ResourceEnum) error {
 }
 
 func (aws *Aws) Delete(resources kftypes.ResourceEnum) error {
+	setAwsPluginDefaultsErr := aws.setAwsPluginDefaults()
+	if setAwsPluginDefaultsErr != nil {
+		return &kfapis.KfError{
+			Code: setAwsPluginDefaultsErr.(*kfapis.KfError).Code,
+			Message: fmt.Sprintf("aws set aws plugin defaults Error %v",
+				setAwsPluginDefaultsErr.(*kfapis.KfError).Message),
+		}
+	}
+
 	awsPlatformUninstallError := aws.uninstallAwsPlatform()
 	if awsPlatformUninstallError != nil {
 		return &kfapis.KfError{
@@ -548,14 +640,14 @@ func (aws *Aws) Delete(resources kftypes.ResourceEnum) error {
 
 // writeConfigFile writes KfDef to app.yaml
 func (aws *Aws) writeConfigFile() error {
-	buf, bufErr := yaml.Marshal(aws.KfDef)
+	buf, bufErr := yaml.Marshal(aws.kfDef)
 	if bufErr != nil {
 		return &kfapis.KfError{
 			Code:    int(kfapis.INVALID_ARGUMENT),
 			Message: fmt.Sprintf("AWS marshaling error: %v", bufErr),
 		}
 	}
-	cfgFilePath := filepath.Join(aws.Spec.AppDir, kftypes.KfConfigFile)
+	cfgFilePath := filepath.Join(aws.kfDef.Spec.AppDir, kftypes.KfConfigFile)
 	cfgFilePathErr := ioutil.WriteFile(cfgFilePath, buf, 0644)
 	if cfgFilePathErr != nil {
 		return &kfapis.KfError{
@@ -575,7 +667,7 @@ func (aws *Aws) writeFeatureConfig(featureConfig map[string]interface{}) error {
 			Message: fmt.Sprintf("AWS marshaling error: %v", bufErr),
 		}
 	}
-	featureCfgFilePath := filepath.Join(aws.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR, CLUSTER_FEATURE_CONFIG_FILE)
+	featureCfgFilePath := filepath.Join(aws.kfDef.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR, CLUSTER_FEATURE_CONFIG_FILE)
 	featureCfgFilePathErr := ioutil.WriteFile(featureCfgFilePath, buf, 0644)
 	if featureCfgFilePathErr != nil {
 		return &kfapis.KfError{
@@ -587,7 +679,7 @@ func (aws *Aws) writeFeatureConfig(featureConfig map[string]interface{}) error {
 }
 
 func (aws *Aws) getFeatureConfig() (map[string]interface{}, error) {
-	configPath := filepath.Join(aws.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR, CLUSTER_FEATURE_CONFIG_FILE)
+	configPath := filepath.Join(aws.kfDef.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR, CLUSTER_FEATURE_CONFIG_FILE)
 	log.Infof("Reading config file: %v", configPath)
 
 	configBuf, bufErr := ioutil.ReadFile(configPath)
@@ -626,8 +718,14 @@ func (aws *Aws) uninstallAwsPlatform() error {
 		}
 	}
 
+	awsPluginSpec, err := aws.GetPluginSpec()
+
+	if err != nil {
+		return err
+	}
+
 	// Detach IAM Policies
-	for _, iamRole := range aws.Spec.Roles {
+	for _, iamRole := range awsPluginSpec.Roles {
 		aws.deleteIamRolePolicy(iamRole, "iam_alb_ingress_policy")
 		aws.deleteIamRolePolicy(iamRole, "iam_csi_fsx_policy")
 
@@ -639,7 +737,7 @@ func (aws *Aws) uninstallAwsPlatform() error {
 	// Delete cluster if it's a managed cluster created by kfctl
 	if config["managed_cluster"] == true {
 		log.Infoln("Start to delete eks cluster. Please wait for 5 mins...")
-		clusterConfigFile := filepath.Join(aws.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR, CLUSTER_CONFIG_FILE)
+		clusterConfigFile := filepath.Join(aws.kfDef.Spec.AppDir, KUBEFLOW_AWS_INFRA_DIR, CLUSTER_CONFIG_FILE)
 		output, err := exec.Command("eksctl", "delete", "cluster", "--config-file="+clusterConfigFile).Output()
 		log.Infoln("Please go to aws console to check CloudFormation status and double make sure your cluster has been shutdown.")
 		if err != nil {
